[CDEC-457] Reunite some orphans from Pos.Core.Genesis.Canonical
