+++ binary/cardano-sl-binary.cabal
@@ -27,6 +27,7 @@ library
   build-depends:       base
                      , binary
                      , bytestring
+                     , canonical-json
                      , cardano-sl-util
                      , cborg
                      , cereal
