+++ core/src/Pos/Core/Common/Coeff.hs
@@ -7,8 +7,11 @@ import           Universum
 import           Data.Fixed (Fixed (..), Nano, showFixed)
 import           Data.SafeCopy (base, deriveSafeCopySimple)
 import qualified Formatting.Buildable as Buildable
+import           Text.JSON.Canonical (FromJSON (..), ReportSchemaErrors,
+                     ToJSON (..))
 
 import           Pos.Binary.Class (Bi (..))
+import           Pos.Core.Genesis.Canonical ()
 
 -- | A fractional coefficient of fixed precision.
 newtype Coeff = Coeff Nano
@@ -22,5 +25,11 @@ instance Bi Coeff where
     decode = Coeff <$> decode @Nano
 
 instance Hashable Coeff
+
+instance Monad m => ToJSON m Coeff where
+    toJSON (Coeff (MkFixed integer)) = toJSON @_ @Integer integer
+
+instance ReportSchemaErrors m => FromJSON m Coeff where
+    fromJSON = fmap (Coeff . MkFixed) . fromJSON @_ @Integer
 
 deriveSafeCopySimple 0 'base ''Coeff
