+++ core/src/Pos/Core/Common/SharedSeed.hs
@@ -8,11 +8,17 @@ import qualified Data.ByteString as BS (pack, zipWith)
 import qualified Data.ByteString.Char8 as BSC (pack)
 import           Data.SafeCopy (base, deriveSafeCopySimple)
 import qualified Data.Semigroup (Semigroup (..))
+import           Formatting (formatToString)
 import qualified Formatting.Buildable as Buildable
-import           Serokell.Util.Base16 (formatBase16)
+import           Serokell.Util.Base16 (base16F, formatBase16)
+import qualified Serokell.Util.Base16 as B16
+import           Text.JSON.Canonical (FromJSON (..), JSValue (..),
+                     ReportSchemaErrors, ToJSON (..))
 
 import           Pos.Binary.Class (Cons (..), Field (..), deriveSimpleBi)
 import           Pos.Core.Constants (sharedSeedLength)
+import           Pos.Core.Genesis.Canonical ()
+import           Pos.Util.Json.Parse (tryParseString)
 
 -- | This is a shared seed used for follow-the-satoshi. This seed is
 -- randomly generated by each party and eventually they agree on the
@@ -32,6 +38,12 @@ instance Monoid SharedSeed where
     mempty = SharedSeed $ BSC.pack $ replicate sharedSeedLength '\NUL'
     mappend = (Data.Semigroup.<>)
     mconcat = foldl' mappend mempty
+
+instance Monad m => ToJSON m SharedSeed where
+    toJSON (SharedSeed seed) = pure . JSString . formatToString base16F $ seed
+
+instance ReportSchemaErrors m => FromJSON m SharedSeed where
+    fromJSON = fmap SharedSeed . tryParseString B16.decode
 
 deriveSimpleBi ''SharedSeed [
     Cons 'SharedSeed [
