+++ core/src/Pos/Core/Delegation/HeavyDlgIndex.hs
@@ -10,6 +10,8 @@ import           Universum
 import           Data.SafeCopy (SafeCopy (..), contain, safeGet, safePut)
 import           Formatting (bprint, build)
 import qualified Formatting.Buildable
+import           Text.JSON.Canonical (FromJSON (..), Int54, JSValue (..),
+                     ReportSchemaErrors, ToJSON (..), fromJSField, mkObject)
 
 import           Pos.Binary.Class (Bi (..))
 import           Pos.Core.Slotting (EpochIndex)
@@ -42,6 +44,25 @@ type ProxySigHeavy a = ProxySignature HeavyDlgIndex a
 
 -- | Heavy delegation PSK.
 type ProxySKHeavy = ProxySecretKey HeavyDlgIndex
+
+instance Monad m => ToJSON m ProxySKHeavy where
+    toJSON psk =
+        -- omega is encoded as a number, because in genesis we always
+        -- set it to 0.
+        mkObject
+            [ ("omega", pure (JSNum . fromIntegral . getHeavyDlgIndex $ pskOmega psk))
+            , ("issuerPk", toJSON $ pskIssuerPk psk)
+            , ("delegatePk", toJSON $ pskDelegatePk psk)
+            , ("cert", toJSON $ pskCert psk)
+            ]
+
+instance ReportSchemaErrors m => FromJSON m ProxySKHeavy where
+    fromJSON obj = do
+        pskOmega <- HeavyDlgIndex . fromIntegral @Int54 <$> fromJSField obj "omega"
+        pskIssuerPk <- fromJSField obj "issuerPk"
+        pskDelegatePk <- fromJSField obj "delegatePk"
+        pskCert <- fromJSField obj "cert"
+        pure UnsafeProxySecretKey{..}
 
 -- | Heavyweight PSK with real leader public key (because heavyweight
 -- psks have redelegation feature, so pskIssuerPk hPsk /= leader in
