+++ core/src/Pos/Core/Genesis/VssCertificatesMap.hs
@@ -9,7 +9,9 @@ import           Data.Semigroup (Semigroup)
 import           Formatting (bprint, (%))
 import qualified Formatting.Buildable as Buildable
 import           Serokell.Util (mapJson)
+import           Text.JSON.Canonical (ToJSON (..))
 
+import           Pos.Core.Genesis.Canonical ()
 import           Pos.Core.Ssc (VssCertificatesMap, getVssCertificatesMap)
 
 -- | Predefined balances of non avvm entries.
@@ -25,3 +27,6 @@ newtype GenesisVssCertificatesMap = GenesisVssCertificatesMap
 instance Buildable GenesisVssCertificatesMap where
     build (GenesisVssCertificatesMap m) =
         bprint ("GenesisVssCertificatesMap: "%mapJson) (getVssCertificatesMap m)
+
+instance Monad m => ToJSON m GenesisVssCertificatesMap where
+    toJSON = toJSON . getGenesisVssCertificatesMap
