+++ core/src/Pos/Core/Slotting/EpochIndex.hs
@@ -12,8 +12,11 @@ import           Data.SafeCopy (base, deriveSafeCopySimple)
 import           Formatting (bprint, int, (%))
 import qualified Formatting.Buildable as Buildable
 import           Servant.API (FromHttpApiData)
+import           Text.JSON.Canonical (FromJSON (..), ReportSchemaErrors,
+                     ToJSON (..))
 
 import           Pos.Binary.Class (Bi (..))
+import           Pos.Core.Genesis.Canonical ()
 import           Pos.Util.Some (Some, liftLensSome)
 
 -- | Index of epoch.
@@ -39,6 +42,14 @@ instance HasEpochIndex (Some HasEpochIndex) where
 instance (HasEpochIndex a, HasEpochIndex b) =>
          HasEpochIndex (Either a b) where
     epochIndexL = choosing epochIndexL epochIndexL
+
+-- Note that it will be encoded as string, because 'EpochIndex'
+-- doesn't necessary fit into JS number.
+instance Monad m => ToJSON m EpochIndex where
+    toJSON = toJSON . getEpochIndex
+
+instance ReportSchemaErrors m => FromJSON m EpochIndex where
+    fromJSON = fmap EpochIndex . fromJSON
 
 -- | Bootstrap era is ongoing until stakes are unlocked. The reward era starts
 -- from the epoch specified as the epoch that unlocks stakes:
