+++ core/src/Pos/Core/Ssc/VssCertificate.hs
@@ -22,9 +22,12 @@ import           Data.SafeCopy (base, deriveSafeCopySimple)
 import           Formatting (bprint, build, int, (%))
 import qualified Formatting.Buildable as Buildable
 import           Pos.Core.Common (StakeholderId, addressHash)
+import           Text.JSON.Canonical (FromJSON (..), Int54, JSValue (..),
+                     ReportSchemaErrors, ToJSON (..), fromJSField, mkObject)
 
 import           Pos.Binary.Class (AsBinary, Bi (..), encodeListLen,
                      enforceSize)
+import           Pos.Core.Genesis.Canonical ()
 import           Pos.Core.Slotting (EpochIndex)
 import           Pos.Crypto (ProtocolMagic, PublicKey, SecretKey,
                      SignTag (SignVssCert), Signature, VssPublicKey, checkSig,
@@ -86,6 +89,28 @@ instance Bi VssCertificate where
         sig <- decode
         sky <- decode
         pure $ UnsafeVssCertificate key epo sig sky
+
+instance Monad m => ToJSON m VssCertificate where
+    toJSON vc =
+        mkObject
+            [ ("vssKey", toJSON (vcVssKey vc))
+            , ("expiryEpoch", pure (JSNum . fromIntegral $ vcExpiryEpoch vc))
+            , ("signature", toJSON (vcSignature vc))
+            , ("signingKey", toJSON (vcSigningKey vc))
+            ]
+
+instance (ReportSchemaErrors m) => FromJSON m VssCertificate where
+    fromJSON obj = do
+        vssKey <- fromJSField obj "vssKey"
+        expiryEpoch <- fromIntegral @Int54 <$> fromJSField obj "expiryEpoch"
+        signature <- fromJSField obj "signature"
+        signingKey <- fromJSField obj "signingKey"
+        return $ UnsafeVssCertificate
+            { vcVssKey      = vssKey
+            , vcExpiryEpoch = expiryEpoch
+            , vcSignature   = signature
+            , vcSigningKey  = signingKey
+            }
 
 -- | Make VssCertificate valid up to given epoch using 'SecretKey' to sign
 -- data.
