+++ crypto/Pos/Crypto/Signing/Types/Redeem.hs
@@ -10,7 +10,10 @@ module Pos.Crypto.Signing.Types.Redeem
        , redeemToPublic
        ) where
 
+import           Universum
+
 import           Control.Exception.Safe (Exception (..))
+import           Control.Lens (_Left)
 import qualified Crypto.Sign.Ed25519 as Ed25519
 import           Data.Aeson.TH (defaultOptions, deriveJSON)
 import qualified Data.ByteString as BS
@@ -19,14 +22,17 @@ import           Data.SafeCopy (SafeCopy (..), base, contain,
                      deriveSafeCopySimple, safeGet, safePut)
 import qualified Data.Text as T
 import qualified Data.Text.Lazy.Builder as Builder (fromText)
-import           Formatting (Format, bprint, fitLeft, later, (%), (%.))
+import           Formatting (Format, bprint, fitLeft, formatToString, later,
+                     (%), (%.))
 import qualified Formatting.Buildable as B
 import           Serokell.Util.Base64 (formatBase64)
 import qualified Serokell.Util.Base64 as B64
-import           Universum
+import           Text.JSON.Canonical (FromObjectKey (..), JSValue (..),
+                     ReportSchemaErrors, ToObjectKey (..))
 
 import           Pos.Binary.Class (Bi)
 import           Pos.Crypto.Orphans ()
+import           Pos.Util.Json.Parse (tryParseString)
 
 ----------------------------------------------------------------------------
 -- PK/SK and formatters
@@ -36,19 +42,20 @@ import           Pos.Crypto.Orphans ()
 newtype RedeemPublicKey = RedeemPublicKey Ed25519.PublicKey
     deriving (Eq, Ord, Show, Generic, NFData, Hashable, Typeable)
 
-deriveJSON defaultOptions ''RedeemPublicKey
+instance Monad m => ToObjectKey m RedeemPublicKey where
+    toObjectKey = pure . formatToString redeemPkB64UrlF
 
-deriving instance Bi RedeemPublicKey
-
-deriveSafeCopySimple 0 'base ''RedeemPublicKey
+instance ReportSchemaErrors m => FromObjectKey m RedeemPublicKey where
+    fromObjectKey =
+        fmap Just .
+        tryParseString (over _Left pretty . fromAvvmPk) .
+        JSString
 
 -- | Wrapper around 'Ed25519.SecretKey'.
 newtype RedeemSecretKey = RedeemSecretKey Ed25519.SecretKey
     deriving (Eq, Ord, Show, Generic, NFData, Hashable)
 
 deriving instance Bi RedeemSecretKey
-
-deriveSafeCopySimple 0 'base ''RedeemSecretKey
 
 redeemPkB64F :: Format r (RedeemPublicKey -> r)
 redeemPkB64F =
@@ -73,6 +80,8 @@ instance B.Buildable RedeemPublicKey where
 instance B.Buildable RedeemSecretKey where
     build = bprint ("redeem_sec_of_pk:"%redeemPkB64F) . redeemToPublic
 
+deriving instance Bi RedeemPublicKey
+
 ----------------------------------------------------------------------------
 -- Redeem signatures
 ----------------------------------------------------------------------------
@@ -83,8 +92,6 @@ newtype RedeemSignature a = RedeemSignature Ed25519.Signature
 
 instance B.Buildable (RedeemSignature a) where
     build _ = "<redeem signature>"
-
-deriveJSON defaultOptions ''RedeemSignature
 
 deriving instance Typeable a => Bi (RedeemSignature a)
 
@@ -132,3 +139,16 @@ redeemPkBuild bs
         "consRedeemPk: failed to form pk, wrong bs length: " <> show (BS.length bs) <>
         ", when should be 32"
     | otherwise = RedeemPublicKey $ Ed25519.PublicKey $ bs
+
+----------------------------------------------------------------------------
+-- Helpers
+----------------------------------------------------------------------------
+
+-- These are *not* orphan instances, these types are defined in this file.
+-- However these need to be defined here to avoid TemplateHaskell compile
+-- phase errors.
+
+deriveSafeCopySimple 0 'base ''RedeemSecretKey
+deriveSafeCopySimple 0 'base ''RedeemPublicKey
+deriveJSON defaultOptions ''RedeemPublicKey
+deriveJSON defaultOptions ''RedeemSignature
