+++ crypto/Pos/Crypto/Signing/Types/Signing.hs
@@ -34,6 +34,8 @@ module Pos.Crypto.Signing.Types.Signing
        , isSelfSignedPsk
        ) where
 
+import           Universum hiding (show)
+
 import qualified Cardano.Crypto.Wallet as CC
 import qualified Codec.CBOR.Decoding as D
 import qualified Codec.CBOR.Encoding as E
@@ -44,18 +46,20 @@ import           Data.Hashable (Hashable)
 import           Data.SafeCopy (SafeCopy (..), base, contain,
                      deriveSafeCopySimple, safeGet, safePut)
 import           Data.Text.Lazy.Builder (Builder)
-import           Formatting (Format, bprint, build, fitLeft, later, sformat,
-                     (%), (%.))
+import           Formatting (Format, bprint, build, fitLeft, formatToString,
+                     later, sformat, (%), (%.))
 import qualified Formatting.Buildable as B
 import           Prelude (show)
 import qualified Serokell.Util.Base16 as B16
 import qualified Serokell.Util.Base64 as Base64 (decode, formatBase64)
-import           Universum hiding (show)
+import           Text.JSON.Canonical (JSValue (..), ReportSchemaErrors)
+import qualified Text.JSON.Canonical as TJC (FromJSON (..), ToJSON (..))
 
 import           Pos.Binary.Class (Bi (..), encodeListLen, enforceSize)
 import qualified Pos.Binary.Class as Bi
 import           Pos.Crypto.Hashing (hash)
 import           Pos.Crypto.Orphans ()
+import           Pos.Util.Json.Parse (tryParseString)
 import           Pos.Util.Util (cerealError, toAesonError, toCborError)
 
 ----------------------------------------------------------------------------
@@ -79,6 +83,12 @@ instance ToJSON PublicKey where
 
 instance FromJSON PublicKey where
     parseJSON v = parseJSON v >>= toAesonError . fmsg "PublicKey" . parseFullPublicKey
+
+instance Monad m => TJC.ToJSON m PublicKey where
+    toJSON = pure . JSString . formatToString fullPublicKeyF
+
+instance ReportSchemaErrors m => TJC.FromJSON m PublicKey where
+    fromJSON = tryParseString parseFullPublicKey
 
 encodeXPub :: CC.XPub -> E.Encoding
 encodeXPub a = encode $ CC.unXPub a
@@ -163,6 +173,12 @@ instance FromJSON (Signature w) where
 instance ToJSON (Signature w) where
     toJSON = toJSON . sformat fullSignatureHexF
 
+instance Monad m => TJC.ToJSON m (Signature w) where
+    toJSON = pure . JSString . formatToString fullSignatureHexF
+
+instance (Typeable x, ReportSchemaErrors m) => TJC.FromJSON m (Signature x) where
+    fromJSON = tryParseString parseFullSignature
+
 instance SafeCopy (Signature a) where
     putCopy (Signature sig) = contain $ safePut sig
     getCopy = contain $ Signature <$> safeGet
@@ -228,6 +244,12 @@ instance ToJSON (ProxyCert w) where
 instance FromJSON (ProxyCert w) where
     parseJSON v = parseJSON v >>= toAesonError . fmsg "Signature" . parseFullProxyCert
 
+instance Monad m => TJC.ToJSON m (ProxyCert w) where
+    toJSON = pure . JSString . formatToString fullProxyCertHexF
+
+instance (Typeable w, ReportSchemaErrors m) => TJC.FromJSON m (ProxyCert w) where
+    fromJSON = tryParseString parseFullProxyCert
+
 instance Typeable w => Bi (ProxyCert w) where
     encode (ProxyCert a) = encodeXSignature a
     decode = fmap ProxyCert decodeXSignature
@@ -262,8 +284,6 @@ instance Hashable w => Hashable (ProxySecretKey w)
 instance (B.Buildable w) => B.Buildable (ProxySecretKey w) where
     build (UnsafeProxySecretKey w iPk dPk _) =
         bprint ("ProxySk { w = "%build%", iPk = "%build%", dPk = "%build%" }") w iPk dPk
-
-deriveJSON defaultOptions ''ProxySecretKey
 
 instance Bi w => Bi (ProxySecretKey w) where
     encode UnsafeProxySecretKey{..} =
@@ -319,11 +339,11 @@ instance SafeCopy w => SafeCopy (ProxySignature w a) where
 isSelfSignedPsk :: ProxySecretKey w -> Bool
 isSelfSignedPsk psk = pskIssuerPk psk == pskDelegatePk psk
 
-
 -- These are *not* orphan instances, these types are defined in this file.
 -- However these need to be defined here to avoid TemplateHaskell compile
 -- phase errors.
 
+deriveJSON defaultOptions ''ProxySecretKey
 deriveSafeCopySimple 0 'base ''PublicKey
 deriveSafeCopySimple 0 'base ''SecretKey
 deriveSafeCopySimple 0 'base ''ProxySecretKey
